+++ example/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import {StyleSheet, View} from 'react-native';
 import PopoverExample from './PopoverExample';
 
 const styles = StyleSheet.create({
@@ -9,7 +9,13 @@ const styles = StyleSheet.create({
     backgroundColor: '#c2ffd2',
     flexDirection: 'row',
     flexWrap: 'wrap',
-  }
+  },
+  button: {
+    position: 'absolute',
+    bottom: 0,
+    left: 0,
+    right: 0,
+  },
 });
 
 const cenralPopoverStyles = StyleSheet.create({
@@ -22,27 +28,75 @@ const cenralPopoverStyles = StyleSheet.create({
     borderTopColor: 'pink',
   },
   backgroundStyle: {
-    backgroundColor: 'rgba(0, 0, 255, 0.5)'
+    backgroundColor: 'rgba(0, 0, 255, 0.5)',
   },
 });
 
 const examples = [
-  {icon: 'arrow-top-left', text: 'I am top-left popover', alignItems: 'flex-start', justifyContent: 'flex-start'},
-  {icon: 'arrow-up', text: 'I am top-middle popover', alignItems: 'center', justifyContent: 'flex-start'},
-  {icon: 'arrow-top-right', text: 'I am top-right popover', alignItems: 'flex-end', justifyContent: 'flex-start'},
+  {
+    icon: 'arrow-top-left',
+    text: 'I am top-left popover',
+    alignItems: 'flex-start',
+    justifyContent: 'flex-start',
+  },
+  {
+    icon: 'arrow-up',
+    text: 'I am top-middle popover',
+    alignItems: 'center',
+    justifyContent: 'flex-start',
+  },
+  {
+    icon: 'arrow-top-right',
+    text: 'I am top-right popover',
+    alignItems: 'flex-end',
+    justifyContent: 'flex-start',
+  },
 
-  {icon: 'arrow-left', text: 'I am middle-left popover', alignItems: 'flex-start', justifyContent: 'center'},
-  {icon: 'image-filter-center-focus', text: 'I am middle-middle popover.\nI am stylish!', alignItems: 'center', justifyContent: 'center', popoverStyles: cenralPopoverStyles},
-  {icon: 'arrow-right', text: 'I am middle-right popover', alignItems: 'flex-end', justifyContent: 'center'},
+  {
+    icon: 'arrow-left',
+    text: 'I am middle-left popover',
+    alignItems: 'flex-start',
+    justifyContent: 'center',
+  },
+  {
+    icon: 'image-filter-center-focus',
+    text: 'I am middle-middle popover.\nI am stylish!',
+    alignItems: 'center',
+    justifyContent: 'center',
+    popoverStyles: cenralPopoverStyles,
+  },
+  {
+    icon: 'arrow-right',
+    text: 'I am middle-right popover',
+    alignItems: 'flex-end',
+    justifyContent: 'center',
+  },
 
-  {icon: 'arrow-bottom-left', text: 'I am bottom-left popover', alignItems: 'flex-start', justifyContent: 'flex-end'},
-  {icon: 'arrow-down', text: 'I am bottom-middle popover', alignItems: 'center', justifyContent: 'flex-end'},
-  {icon: 'arrow-bottom-right', text: 'I am bottom-right popover', alignItems: 'flex-end', justifyContent: 'flex-end'},
+  {
+    icon: 'arrow-bottom-left',
+    text: 'I am bottom-left popover',
+    alignItems: 'flex-start',
+    justifyContent: 'flex-end',
+  },
+  {
+    icon: 'arrow-down',
+    text: 'I am bottom-middle popover',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+  },
+  {
+    icon: 'arrow-bottom-right',
+    text: 'I am bottom-right popover',
+    alignItems: 'flex-end',
+    justifyContent: 'flex-end',
+  },
 ];
 
 const App = () => (
   <View style={styles.app}>
-    { examples.map(example => <PopoverExample key={example.icon} {...example} />)}
+    {examples.map((example) => (
+      <PopoverExample key={example.icon} {...example} />
+    ))}
   </View>
 );