+++ src/PopoverGeometry.ts
@@ -1,3 +1,5 @@
+import {I18nManager} from 'react-native';
+
 export interface Point {
   x: number;
   y: number;
@@ -15,7 +17,7 @@ export interface Size {
   height: number;
 }
 
-export type Placement = 'top' | 'right' | 'bottom' | 'left';
+export type Placement = 'top' | 'end' | 'bottom' | 'start';
 
 export interface Geometry {
   origin: Point;
@@ -23,96 +25,198 @@ export interface Geometry {
   placement: Placement;
 }
 
-type ComputeGeometry = (displayArea: Rect, fromRect: Rect, contentSize: Size, arrowSize: Size) => Geometry;
+type ComputeGeometry = (
+  displayArea: Rect,
+  fromRect: Rect,
+  contentSize: Size,
+  arrowSize: Size,
+) => Geometry;
 
-export const computeGeometry =
-  (contentSize: Size, placement: Placement | 'auto', fromRect: Rect, displayArea: Rect, arrowSize: Size): Geometry => {
+export const computeGeometry = (
+  contentSize: Size,
+  placement: Placement | 'auto',
+  fromRect: Rect,
+  displayArea: Rect,
+  arrowSize: Size,
+): Geometry => {
   const effectiveArrowSize = getArrowSize(arrowSize, placement);
 
   switch (placement) {
     case 'top':
-      return computeTopGeometry(displayArea, fromRect, contentSize, effectiveArrowSize);
+      return computeTopGeometry(
+        displayArea,
+        fromRect,
+        contentSize,
+        effectiveArrowSize,
+      );
     case 'bottom':
-      return computeBottomGeometry(displayArea, fromRect, contentSize, effectiveArrowSize);
-    case 'left':
-      return computeLeftGeometry(displayArea, fromRect, contentSize, effectiveArrowSize);
-    case 'right':
-      return computeRightGeometry(displayArea, fromRect, contentSize, effectiveArrowSize);
+      return computeBottomGeometry(
+        displayArea,
+        fromRect,
+        contentSize,
+        effectiveArrowSize,
+      );
+    case 'start':
+      return computeStartGeometry(
+        displayArea,
+        fromRect,
+        contentSize,
+        effectiveArrowSize,
+      );
+    case 'end':
+      return computeEndGeometry(
+        displayArea,
+        fromRect,
+        contentSize,
+        effectiveArrowSize,
+      );
     default:
-      return computeAutoGeometry(displayArea, fromRect, contentSize, effectiveArrowSize);
+      return computeAutoGeometry(
+        displayArea,
+        fromRect,
+        contentSize,
+        effectiveArrowSize,
+      );
   }
 };
 
 const getArrowSize = (size: Size, placement: Placement | 'auto'): Size => {
-  if (placement === 'left' || placement === 'right') {
-    return { width: size.height, height: size.width };
+  if (placement === 'start' || placement === 'end') {
+    return {width: size.height, height: size.width};
   }
   return size;
 };
 
-const computeTopGeometry: ComputeGeometry = (displayArea, fromRect, contentSize, arrowSize) => {
+const computeTopGeometry: ComputeGeometry = (
+  displayArea,
+  fromRect,
+  contentSize,
+  arrowSize,
+) => {
   const origin = {
     x: Math.min(
       displayArea.x + displayArea.width - contentSize.width,
-      Math.max(displayArea.x, fromRect.x + (fromRect.width - contentSize.width) / 2),
+      Math.max(
+        displayArea.x,
+        fromRect.x +
+          (fromRect.width - contentSize.width) / 2 -
+          (I18nManager.isRTL ? fromRect.width : 0),
+      ),
     ),
     y: fromRect.y - contentSize.height - arrowSize.height,
   };
 
-  const anchor = { x: fromRect.x + fromRect.width / 2, y: fromRect.y };
+  const anchor = {
+    x: fromRect.x + (fromRect.width / 2) * (I18nManager.isRTL ? -1 : 1),
+    y: fromRect.y,
+  };
 
-  return { origin, anchor, placement: 'top' };
+  return {origin, anchor, placement: 'top'};
 };
 
-const computeBottomGeometry: ComputeGeometry = (displayArea, fromRect, contentSize, arrowSize) => {
+const computeBottomGeometry: ComputeGeometry = (
+  displayArea,
+  fromRect,
+  contentSize,
+  arrowSize,
+) => {
   const origin = {
     x: Math.min(
       displayArea.x + displayArea.width - contentSize.width,
-      Math.max(displayArea.x, fromRect.x + (fromRect.width - contentSize.width) / 2),
+      Math.max(
+        displayArea.x,
+        fromRect.x +
+          (fromRect.width - contentSize.width) / 2 -
+          (I18nManager.isRTL ? fromRect.width : 0),
+      ),
     ),
     y: fromRect.y + fromRect.height + arrowSize.height,
   };
 
-  const anchor = { x: fromRect.x + fromRect.width / 2, y: fromRect.y + fromRect.height };
+  const anchor = {
+    x: fromRect.x + (fromRect.width / 2) * (I18nManager.isRTL ? -1 : 1),
+    y: fromRect.y + fromRect.height,
+  };
 
-  return { origin, anchor, placement: 'bottom' };
+  return {origin, anchor, placement: 'bottom'};
 };
 
-const computeLeftGeometry: ComputeGeometry = (displayArea, fromRect, contentSize, arrowSize) => {
+const computeStartGeometry: ComputeGeometry = (
+  displayArea,
+  fromRect,
+  contentSize,
+  arrowSize,
+) => {
   const origin = {
-    x: fromRect.x - contentSize.width - arrowSize.width,
+    x:
+      fromRect.x -
+      contentSize.width -
+      arrowSize.width -
+      (I18nManager.isRTL ? fromRect.width : 0),
     y: Math.min(
       displayArea.y + displayArea.height - contentSize.height,
-      Math.max(displayArea.y, fromRect.y + (fromRect.height - contentSize.height) / 2),
+      Math.max(
+        displayArea.y,
+        fromRect.y + fromRect.height - contentSize.height,
+      ),
     ),
   };
 
-  const anchor = { x: fromRect.x, y: fromRect.y + fromRect.height / 2 };
+  const anchor = {
+    x: fromRect.x - (I18nManager.isRTL ? fromRect.width : 0),
+    y: fromRect.y + fromRect.height - contentSize.height / 2,
+  };
 
-  return { origin, anchor, placement: 'left' };
+  return {origin, anchor, placement: 'start'};
 };
 
-const computeRightGeometry: ComputeGeometry = (displayArea, fromRect, contentSize, arrowSize) => {
+const computeEndGeometry: ComputeGeometry = (
+  displayArea,
+  fromRect,
+  contentSize,
+  arrowSize,
+) => {
   const origin = {
-    x: fromRect.x + fromRect.width + arrowSize.width,
+    x:
+      fromRect.x +
+      fromRect.width +
+      arrowSize.width -
+      (I18nManager.isRTL ? fromRect.width : 0),
     y: Math.min(
       displayArea.y + displayArea.height - contentSize.height,
-      Math.max(displayArea.y, fromRect.y + (fromRect.height - contentSize.height) / 2),
+      Math.max(
+        displayArea.y,
+        fromRect.y + fromRect.height - contentSize.height,
+      ),
     ),
   };
 
-  const anchor = { x: fromRect.x + fromRect.width, y: fromRect.y + fromRect.height / 2 };
+  const anchor = {
+    x: fromRect.x + (I18nManager.isRTL ? 0 : fromRect.width),
+    y: fromRect.y + fromRect.height - contentSize.height / 2,
+  };
 
-  return { origin, anchor, placement: 'right' };
+  return {origin, anchor, placement: 'end'};
 };
 
-const computeAutoGeometry = (displayArea: Rect, fromRect: Rect, contentSize: Size, arrowSize: Size): Geometry => {
+const computeAutoGeometry = (
+  displayArea: Rect,
+  fromRect: Rect,
+  contentSize: Size,
+  arrowSize: Size,
+): Geometry => {
   let geom: Geometry | null = null;
-  const placements: Placement[] = ['left', 'top', 'right', 'bottom'];
+  const placements: Placement[] = ['start', 'end', 'top', 'bottom'];
   for (let i = 0; i < 4; i += 1) {
     const placement = placements[i];
-    geom = computeGeometry(contentSize, placement, fromRect, displayArea, arrowSize);
-    const { origin } = geom;
+    geom = computeGeometry(
+      contentSize,
+      placement,
+      fromRect,
+      displayArea,
+      arrowSize,
+    );
+    const {origin} = geom;
 
     if (
       origin.x >= displayArea.x &&