+++ src/PopoverTouchable.tsx
@@ -1,7 +1,11 @@
-import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import { Dimensions, findNodeHandle, MeasureOnSuccessCallback, NativeModules, View } from 'react-native';
-import { Rect } from './PopoverGeometry';
+import {
+  Dimensions,
+  findNodeHandle,
+  MeasureOnSuccessCallback,
+  NativeModules,
+} from 'react-native';
+import {Rect} from './PopoverGeometry';
 
 export interface Props {
   onPopoverDisplayed?: () => any;
@@ -13,16 +17,11 @@ export interface State {
 }
 
 export class PopoverTouchable extends React.PureComponent<Props, State> {
-
-  static propTypes = {
-    onPopoverDisplayed: PropTypes.func,
-  };
-
   constructor(props: Props) {
     super(props);
     this.state = {
       showPopover: false,
-      popoverAnchor: { x: 0, y: 0, width: 0, height: 0 },
+      popoverAnchor: {x: 0, y: 0, width: 0, height: 0},
     };
     // Not yet
     // console.warn('PopoverTouchable is deprecated, please use PopoverController instead');
@@ -56,11 +55,18 @@ export class PopoverTouchable extends React.PureComponent<Props, State> {
     }
   };
 
-  private onTouchableMeasured: MeasureOnSuccessCallback = (x0, y0, width, height, x, y) => {
+  private onTouchableMeasured: MeasureOnSuccessCallback = (
+    x0,
+    y0,
+    width,
+    height,
+    x,
+    y,
+  ) => {
     this.setState(
       {
         showPopover: true,
-        popoverAnchor: { x, y, width, height },
+        popoverAnchor: {x, y, width, height},
       },
       () => {
         if (this.props.onPopoverDisplayed) {
@@ -70,7 +76,7 @@ export class PopoverTouchable extends React.PureComponent<Props, State> {
     );
   };
 
-  private onClosePopover = () => this.setState({ showPopover: false });
+  private onClosePopover = () => this.setState({showPopover: false});
 
   render() {
     const children = React.Children.toArray(this.props.children);
@@ -80,23 +86,21 @@ export class PopoverTouchable extends React.PureComponent<Props, State> {
       children[1] instanceof String ||
       (children[1] as any).type.displayName !== 'Popover'
     ) {
-      throw new Error('Popover touchable must have two children and the second one must be Popover');
+      throw new Error(
+        'Popover touchable must have two children and the second one must be Popover',
+      );
     }
     return (
       <React.Fragment>
-        {
-          React.cloneElement(children[0] as React.ReactElement<any>, {
-            ref: this.setRef,
-            onPress: this.onPress,
-          })
-        }
-        {
-          React.cloneElement(children[1] as React.ReactElement<any>, {
-            visible: this.state.showPopover,
-            onClose: this.onClosePopover,
-            fromRect: this.state.popoverAnchor,
-          })
-        }
+        {React.cloneElement(children[0] as React.ReactElement<any>, {
+          ref: this.setRef,
+          onPress: this.onPress,
+        })}
+        {React.cloneElement(children[1] as React.ReactElement<any>, {
+          visible: this.state.showPopover,
+          onClose: this.onClosePopover,
+          fromRect: this.state.popoverAnchor,
+        })}
       </React.Fragment>
     );
   }