+++ src/index.ts
@@ -1,7 +1,8 @@
-import { Popover } from './Popover';
+import {Popover} from './Popover';
 
 export default Popover;
 
-export { Popover, PopoverProps } from './Popover';
-export { PopoverController, PopoverControllerRenderProps } from './PopoverController';
-export { PopoverTouchable } from './PopoverTouchable';
+export type {PopoverProps} from './Popover';
+export {PopoverController} from './PopoverController';
+export type {PopoverControllerRenderProps} from './PopoverController';
+export {PopoverTouchable} from './PopoverTouchable';