+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change log
 
-### vNext
+### 0.0.12
+
+Fix typescript definitions for styles
+
+### 0.0.11
+
+Update dev deps and fix tsc error 
+
+### 0.0.10
+
+Exposed onDismiss prop
+
+### 0.0.9
+
+Fix ts props/default props 
 
 ### 0.0.7